traffic_density_estimator_service/requirements.txt
grpcio==1.35.0
grpcio-reflection==1.35.0
Pillow==8.1.2
protobuf==3.14.0
torch==1.7.1
torchvision==0.8.2
pyyaml